🎨 背景に関連する要素をBackgroundComponentに整理
